requirements.txt
RPLCD
smbus2
RPi.GPIO
sounddevice
scipy
numpy
requests
spidev
Pillow
